Revert "Migrate getRegisteredSyncRoots"
This reverts commit 8285c0bea08a5dc4327c8db436c7eac914fb5b21.
