+++ packages/addon/include/internal.h
@@ -21,7 +21,6 @@
 #include <virtual_drive/create_folder_placeholder.h>
 #include <virtual_drive/dehydrate_file.h>
 #include <virtual_drive/disconnect_sync_root.h>
-#include <virtual_drive/get_registered_sync_roots.h>
 #include <virtual_drive/get_sync_root_from_path.h>
 #include <virtual_drive/hydrate_file.h>
 #include <virtual_drive/register_sync_root.h>
