+++ packages/addon/include/virtual_drive/connect_sync_root.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <external.h>
+
 inline napi_value connectSyncRoot(napi_env env, napi_callback_info info)
 {
     auto [syncRootPath, fetchDataCallback] =
