+++ packages/addon/include/virtual_drive/convert_to_placeholder.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <external.h>
+
 inline void convert_to_placeholder(const std::wstring& path, const std::wstring& placeholderId)
 {
     auto fileHandle = openFileHandle(path, FILE_READ_ATTRIBUTES | FILE_WRITE_ATTRIBUTES, false);
