+++ packages/addon/include/virtual_drive/create_folder_placeholder.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <external.h>
+
 inline void create_folder_placeholder(const std::wstring& path, const std::wstring& placeholderId, int64_t creationTimeMs, int64_t lastWriteTimeMs)
 {
     if (std::filesystem::exists(path)) {
