+++ packages/addon/include/virtual_drive/dehydrate_file.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <external.h>
+
 inline void dehydrate_file(const std::wstring& path)
 {
     if (std::filesystem::is_directory(path)) {
