+++ packages/addon/include/virtual_drive/get_sync_root_from_path.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <external.h>
+
 struct SyncRoot {
     bool allowPinning;
     bool showSiblingsAsGroup;
