+++ packages/addon/include/virtual_drive/hydrate_file.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <external.h>
+
 inline void hydrate_file(const std::wstring& path)
 {
     if (std::filesystem::is_directory(path)) {
