+++ packages/addon/include/virtual_drive/open_file_handle.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <external.h>
+
 inline winrt::file_handle openFileHandle(const std::wstring& path, DWORD dwDesiredAccess, bool openAsPlaceholder)
 {
     bool isDirectory = std::filesystem::is_directory(path);
