+++ packages/addon/include/virtual_drive/update_placeholder.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <external.h>
+
 inline void update_placeholder(const std::wstring& path, const std::wstring& placeholderId, int64_t size)
 {
     auto fileHandle = openFileHandle(path, FILE_WRITE_ATTRIBUTES, true);
