+++ packages/addon/include/virtual_drive/update_sync_status.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <external.h>
+
 inline void update_sync_status(const std::wstring& path)
 {
     auto fileHandle = openFileHandle(path, FILE_WRITE_ATTRIBUTES, true);
